// src/types/data.ts
export interface Data {
  rodada: string
  data: string
  horario: string
  equipe1: string
  equipe2: string
  placar1: string
  placar2: string
  estadio: string
  local: string
  edicao: string
}

export interface OutdatedData {
  partidas: Partidas[]
  numero: number
}

export interface Partidas {
  visitante: string
  resultado: string
  data_partida: string
  pontuacao_geral_mandante: PontuacaoGeralMandante
  placar_visitante: number
  hora_partida: string
  mandante: string
  placar_mandante: number
  estadio: string
  pontuacao_geral_visitante: PontuacaoGeralVisitante
}

export interface PontuacaoGeralMandante {
  gols_fora_casa: number
  empates_fora_casa: number
  total_jogos: number
  gols_casa: number
  jogos_fora_casa: number
  vitorias_casa: number
  derrotas_casa: number
  total_pontos: number
  empates_casa: number
  pontos_fora_casa: number
  total_gols_sofridos: number
  total_vitorias: number
  vitorias_fora_casa: number
  total_derrotas: number
  pontos_casa: number
  derrotas_fora_casa: number
  total_gols_marcados: number
  jogos_casa: number
  total_empates: number
}

export interface PontuacaoGeralVisitante {
  gols_fora_casa: number
  empates_fora_casa: number
  total_jogos: number
  gols_casa: number
  jogos_fora_casa: number
  vitorias_casa: number
  derrotas_casa: number
  total_pontos: number
  empates_casa: number
  pontos_fora_casa: number
  total_gols_sofridos: number
  total_vitorias: number
  vitorias_fora_casa: number
  total_derrotas: number
  pontos_casa: number
  derrotas_fora_casa: number
  total_gols_marcados: number
  jogos_casa: number
  total_empates: number
}